\name{crab}
\docType{data}
\alias{crab}

\title{Horseshoe Crab Data}

\description{
  A data collected for the purpose of studying what factors determine the number of crab satellites.
}
\usage{data(crabs)}

\format{A data frame of 5 variables and 173 samples.}

\source{From Alan Agresti's "An Introduction to Categorical Data Analysis"}

\keyword{datasets}
