\name{lung}
\docType{data}
\alias{lung}
\alias{lung.alive}
\alias{lung.dead}

\title{Lung cancer data, Boston cohort}
\description{
  Transcriptional profiles of nine genes derived from the BROAD's Boston lung cancer dataset
}
\usage{data(lung)}
\format{A data frame of 9 genes and 64 samples.}
\source{"http://www.broadinstitute.org/gsea/datasets.jsp"}

\keyword{datasets}
